pyproject.toml
[project]
name = "langchain-llm-guide"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain-openai>=1.1.0",
    "langchain[all]>=1.1.0",
    "openai>=2.8.1",
    "python-dotenv>=1.2.1",
]
